requirements.txt
elasticsearch==1.6.0
elasticsearch-dsl==0.0.4
Flask==0.10.1
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
nltk==3.0.3
oauthlib==0.7.2
pbr==1.1.1
python-dateutil==2.4.2
python-instagram==1.3.1
requests==2.7.0
requests-oauthlib==0.5.0
simplejson==3.7.3
six==1.9.0
stevedore==1.5.0
textblob==0.9.1
tweepy==3.3.0
urllib3==1.10.4
virtualenv==13.0.3
virtualenv-clone==0.2.5
virtualenvwrapper==4.6.0
Werkzeug==0.10.4
wheel==0.24.0
